+++ silo-core/test/foundry/Silo/max/maxBorrow/MaxBorrowShares.i.sol
@@ -42,7 +42,7 @@ contract MaxBorrowSharesTest is SiloLittleHelper, Test {
     }
 
     /*
-    forge test -vv --ffi --mt test_maxBorrowShares_withCollateral
+    FOUNDRY_PROFILE=core_test forge test -vv --ffi --mt test_maxBorrowShares_withCollateral
     */
     /// forge-config: core_test.fuzz.runs = 1000
     function test_maxBorrowShares_withCollateral_1token_fuzz(
@@ -62,7 +62,7 @@ contract MaxBorrowSharesTest is SiloLittleHelper, Test {
         uint256 maxBorrowShares = silo1.maxBorrowShares(borrower);
         vm.assume(maxBorrowShares > 0);
 
-        _assertWeCanNotBorrowAboveMax(maxBorrowShares, 1);
+        _assertWeCanNotBorrowAboveMax(maxBorrowShares, 2);
 
         _assertMaxBorrowSharesIsZeroAtTheEnd();
     }
@@ -92,7 +92,7 @@ contract MaxBorrowSharesTest is SiloLittleHelper, Test {
     }
 
     /*
-    forge test -vv --ffi --mt test_maxBorrowShares_withDebt
+    FOUNDRY_PROFILE=core_test forge test -vv --ffi --mt test_maxBorrowShares_withDebt
     */
     /// forge-config: core_test.fuzz.runs = 1000
     function test_maxBorrowShares_withDebt_1token_fuzz(uint128 _collateral, uint128 _liquidity) public {
@@ -116,7 +116,7 @@ contract MaxBorrowSharesTest is SiloLittleHelper, Test {
 
         maxBorrowShares = silo1.maxBorrowShares(borrower);
 
-        _assertWeCanNotBorrowAboveMax(maxBorrowShares, 1);
+        _assertWeCanNotBorrowAboveMax(maxBorrowShares, 2);
         _assertMaxBorrowSharesIsZeroAtTheEnd();
     }
 
@@ -131,6 +131,9 @@ contract MaxBorrowSharesTest is SiloLittleHelper, Test {
         _maxBorrowShares_withInterest_fuzz(_collateral, _liquidity, ISilo.CollateralType.Collateral);
     }
 
+    /*
+    FOUNDRY_PROFILE=core_test forge test -vv --ffi --mt test_maxBorrowShares_withInterest_
+    */
     /// forge-config: core_test.fuzz.runs = 1000
     function test_maxBorrowShares_withInterest_1token_protected_fuzz(
         uint128 _collateral,
@@ -163,7 +166,7 @@ contract MaxBorrowSharesTest is SiloLittleHelper, Test {
         maxBorrowShares = silo1.maxBorrowShares(borrower);
         emit log_named_uint("____ maxBorrowShares", maxBorrowShares);
 
-        _assertWeCanNotBorrowAboveMax(maxBorrowShares, 2);
+        _assertWeCanNotBorrowAboveMax(maxBorrowShares, 3);
         _assertMaxBorrowSharesIsZeroAtTheEnd(1);
     }
 
@@ -179,6 +182,9 @@ contract MaxBorrowSharesTest is SiloLittleHelper, Test {
         _maxBorrowShares_repayWithInterest_fuzz(_collateral, _liquidity, ISilo.CollateralType.Collateral);
     }
 
+    /*
+    FOUNDRY_PROFILE=core_test forge test -vv --ffi --mt test_maxBorrowShares_repayWithInterest_1token_protected_fuzz
+    */
     /// forge-config: core_test.fuzz.runs = 5000
     function test_maxBorrowShares_repayWithInterest_1token_protected_fuzz(
         uint64 _collateral,
@@ -225,7 +231,7 @@ contract MaxBorrowSharesTest is SiloLittleHelper, Test {
         maxBorrowShares = silo1.maxBorrowShares(borrower);
         emit log_named_uint("____ maxBorrowShares", maxBorrowShares);
 
-        _assertWeCanNotBorrowAboveMax(maxBorrowShares, 2);
+        _assertWeCanNotBorrowAboveMax(maxBorrowShares, 3);
         _assertMaxBorrowSharesIsZeroAtTheEnd(1);
     }